Enable toggling of all items display
By default, only the first selected value is displayed in the field, with the number of
additionally selcted values (N) is indicated as "(+N other(s))".

This commit adds the possibility of changing this behavior by setting the displayAllSelected property to true, in which case all selected items will be displayed, comma-separated, with ellipsis if more items are present than fits the component.